+++ docs-markdown/src/test/suite/controllers/quick-pick-menu-controller.test.ts
@@ -1,8 +1,9 @@
 /* eslint-disable @typescript-eslint/no-non-null-assertion */
 import * as chai from 'chai';
 import * as spies from 'chai-spies';
+import * as telemetry from '../../../helper/telemetry';
 import { resolve } from 'path';
-import { commands, QuickPickItem, window, workspace } from 'vscode';
+import { commands, QuickPickItem, window, workspace, ExtensionContext, Uri } from 'vscode';
 import * as boldController from './../../../controllers/bold-controller';
 import * as italicController from '../../../controllers/italic-controller';
 import * as codeController from '../../../controllers/code-controller';
@@ -20,7 +21,6 @@ import * as cleanupController from '../../../controllers/cleanup/cleanup-control
 import * as monikerController from '../../../controllers/moniker-controller';
 import * as yamlController from '../../../controllers/yaml/yaml-controller';
 import { loadDocumentAndGetItReady, sleep, sleepTime } from '../../test.common/common';
-import * as telemetry from '../../../helper/telemetry';
 
 chai.use(spies);
 
@@ -32,6 +32,44 @@ import {
 
 const expect = chai.expect;
 
+interface Subscription {
+	dispose(): any;
+}
+interface EnvironmentalMutator {
+	type: any;
+	value: any;
+}
+const uri = resolve(__dirname, '../../../../../src/test/data/repo/articles/image-controller2.md');
+let environmentalMutator: EnvironmentalMutator;
+let subscriptions: Subscription[];
+const context: ExtensionContext = {
+	globalState: {
+		get: key => {},
+		update: (key, value) => Promise.resolve()
+	},
+	subscriptions,
+	workspaceState: {
+		get: () => {},
+		update: (key, value) => Promise.resolve()
+	},
+	extensionPath: '',
+	asAbsolutePath: relative => '',
+	storagePath: '',
+	globalStoragePath: '',
+	logPath: '',
+	extensionUri: Uri.parse(uri),
+	environmentVariableCollection: {
+		persistent: false,
+		replace: (variable, value) => {},
+		append: (variable, value) => {},
+		prepend: (variable, value) => {},
+		get: variable => environmentalMutator,
+		forEach: () => {},
+		clear: () => {},
+		delete: () => {}
+	}
+};
+
 suite('Quick Pick Menu Controller', () => {
 	suiteSetup(() => {
 		sinon.stub(telemetry, 'sendTelemetryData');
@@ -42,7 +80,7 @@ suite('Quick Pick Menu Controller', () => {
 	});
 	test('quickPickMenuCommand', () => {
 		const controllerCommands = [{ command: markdownQuickPick.name, callback: markdownQuickPick }];
-		expect(quickPickMenuCommand()).to.deep.equal(controllerCommands);
+		expect(quickPickMenuCommand).to.deep.equal(controllerCommands);
 	});
 	test('markdownQuickPick - formatBold', async () => {
 		workspace.getConfiguration = () => {
@@ -67,7 +105,7 @@ suite('Quick Pick Menu Controller', () => {
 		};
 		stubShowQuickPick.onCall(0).resolves(item);
 		const spy = chai.spy.on(boldController, 'formatBold');
-		markdownQuickPick();
+		markdownQuickPick(context);
 		await sleep(sleepTime);
 		expect(spy).to.have.been.called();
 		stubShowQuickPick.restore();
@@ -95,7 +133,7 @@ suite('Quick Pick Menu Controller', () => {
 		};
 		stubShowQuickPick.onCall(0).resolves(item);
 		const spy = chai.spy.on(italicController, 'formatItalic');
-		markdownQuickPick();
+		markdownQuickPick(context);
 		await sleep(sleepTime);
 		expect(spy).to.have.been.called();
 		stubShowQuickPick.restore();
@@ -123,7 +161,7 @@ suite('Quick Pick Menu Controller', () => {
 		};
 		stubShowQuickPick.onCall(0).resolves(item);
 		const spy = chai.spy.on(codeController, 'formatCode');
-		markdownQuickPick();
+		markdownQuickPick(context);
 		await sleep(sleepTime);
 		expect(spy).to.have.been.called();
 		stubShowQuickPick.restore();
@@ -150,8 +188,9 @@ suite('Quick Pick Menu Controller', () => {
 			label: '$(alert) Alert'
 		};
 		stubShowQuickPick.onCall(0).resolves(item);
+		stubShowQuickPick.onCall(1).resolves('some selection');
 		const spy = chai.spy.on(alertController, 'insertAlert');
-		markdownQuickPick();
+		markdownQuickPick(context);
 		await sleep(sleepTime);
 		expect(spy).to.have.been.called();
 		stubShowQuickPick.restore();
@@ -179,7 +218,7 @@ suite('Quick Pick Menu Controller', () => {
 		};
 		stubShowQuickPick.onCall(0).resolves(item);
 		const spy = chai.spy.on(listController, 'insertNumberedList');
-		markdownQuickPick();
+		markdownQuickPick(context);
 		await sleep(sleepTime);
 		expect(spy).to.have.been.called();
 		stubShowQuickPick.restore();
@@ -207,7 +246,7 @@ suite('Quick Pick Menu Controller', () => {
 		};
 		stubShowQuickPick.onCall(0).resolves(item);
 		const spy = chai.spy.on(listController, 'insertBulletedList');
-		markdownQuickPick();
+		markdownQuickPick(context);
 		await sleep(sleepTime);
 		expect(spy).to.have.been.called();
 		stubShowQuickPick.restore();
@@ -235,7 +274,7 @@ suite('Quick Pick Menu Controller', () => {
 		};
 		stubShowQuickPick.onCall(0).resolves(item);
 		const spy = chai.spy.on(tableController, 'insertTable');
-		markdownQuickPick();
+		markdownQuickPick(context);
 		await sleep(sleepTime);
 		expect(spy).to.have.been.called();
 		stubShowQuickPick.restore();
@@ -257,13 +296,18 @@ suite('Quick Pick Menu Controller', () => {
 		);
 		await loadDocumentAndGetItReady(filePath);
 		const stubShowQuickPick = sinon.stub(window, 'showQuickPick');
-		const item: QuickPickItem = {
+		const item1: QuickPickItem = {
 			description: '',
 			label: '$(link) Link'
 		};
-		stubShowQuickPick.onCall(0).resolves(item);
+		const item2: QuickPickItem = {
+			description: '',
+			label: '(foo)'
+		};
+		stubShowQuickPick.onCall(0).resolves(item1);
+		stubShowQuickPick.onCall(1).resolves(item2);
 		const spy = chai.spy.on(linkController, 'pickLinkType');
-		markdownQuickPick();
+		markdownQuickPick(context);
 		await sleep(sleepTime);
 		expect(spy).to.have.been.called();
 		stubShowQuickPick.restore();
@@ -291,7 +335,7 @@ suite('Quick Pick Menu Controller', () => {
 		};
 		stubShowQuickPick.onCall(0).resolves(item);
 		const spy = chai.spy.on(noLocController, 'noLocText');
-		markdownQuickPick();
+		markdownQuickPick(context);
 		await sleep(sleepTime);
 		expect(spy).to.have.been.called();
 		stubShowQuickPick.restore();
@@ -313,13 +357,18 @@ suite('Quick Pick Menu Controller', () => {
 		);
 		await loadDocumentAndGetItReady(filePath);
 		const stubShowQuickPick = sinon.stub(window, 'showQuickPick');
-		const item: QuickPickItem = {
+		const item1: QuickPickItem = {
 			description: '',
 			label: '$(file-media) Image'
 		};
-		stubShowQuickPick.onCall(0).resolves(item);
+		const item2: QuickPickItem = {
+			description: resolve(__dirname, '../../../../../src/test/data/repo/images/'),
+			label: 'test.png'
+		};
+		stubShowQuickPick.onCall(0).resolves(item1);
+		stubShowQuickPick.onCall(1).resolves(item2);
 		const spy = chai.spy.on(imageController, 'pickImageType');
-		markdownQuickPick();
+		markdownQuickPick(context);
 		await sleep(sleepTime);
 		expect(spy).to.have.been.called();
 		stubShowQuickPick.restore();
@@ -347,7 +396,7 @@ suite('Quick Pick Menu Controller', () => {
 		};
 		stubShowQuickPick.onCall(0).resolves(item);
 		const spy = chai.spy.on(includeController, 'insertInclude');
-		markdownQuickPick();
+		markdownQuickPick(context);
 		await sleep(sleepTime);
 		expect(spy).to.have.been.called();
 		stubShowQuickPick.restore();
@@ -375,7 +424,7 @@ suite('Quick Pick Menu Controller', () => {
 		};
 		stubShowQuickPick.onCall(0).resolves(item);
 		const spy = chai.spy.on(snippetController, 'insertSnippet');
-		markdownQuickPick();
+		markdownQuickPick(context);
 		await sleep(sleepTime);
 		expect(spy).to.have.been.called();
 		chai.spy.restore();
@@ -404,7 +453,7 @@ suite('Quick Pick Menu Controller', () => {
 		};
 		stubShowQuickPick.onCall(0).resolves(item);
 		const spy = chai.spy.on(mediaController, 'insertVideo');
-		markdownQuickPick();
+		markdownQuickPick(context);
 		await sleep(sleepTime);
 		expect(spy).to.have.been.called();
 		stubShowQuickPick.restore();
@@ -431,10 +480,12 @@ suite('Quick Pick Menu Controller', () => {
 			label: '$(ellipsis) Columns'
 		};
 		stubShowQuickPick.onCall(0).resolves(item);
+		stubShowQuickPick.onCall(1).resolves('some selection');
 		const spy = chai.spy.on(rowColumnsController, 'insertRowsAndColumns');
-		markdownQuickPick();
+		markdownQuickPick(context);
 		await sleep(sleepTime);
 		expect(spy).to.have.been.called();
+		chai.spy.restore(rowColumnsController, 'insertRowsAndColumns');
 		stubShowQuickPick.restore();
 	});
 	test('markdownQuickPick - applyCleanup', async () => {
@@ -460,7 +511,7 @@ suite('Quick Pick Menu Controller', () => {
 		};
 		stubShowQuickPick.onCall(0).resolves(item);
 		const spy = chai.spy.on(cleanupController, 'applyCleanup');
-		markdownQuickPick();
+		markdownQuickPick(context);
 		await sleep(sleepTime);
 		expect(spy).to.have.been.called();
 		stubShowQuickPick.restore();
@@ -488,7 +539,7 @@ suite('Quick Pick Menu Controller', () => {
 		};
 		stubShowQuickPick.onCall(0).resolves(item);
 		const spy = chai.spy.on(monikerController, 'insertMoniker');
-		markdownQuickPick();
+		markdownQuickPick(context);
 		await sleep(sleepTime);
 		expect(spy).to.have.been.called();
 		stubShowQuickPick.restore();
@@ -515,8 +566,9 @@ suite('Quick Pick Menu Controller', () => {
 			label: '$(note) TOC entry'
 		};
 		stubShowQuickPick.onCall(0).resolves(item);
+		stubShowQuickPick.onCall(1).resolves('some selection');
 		const spy = chai.spy.on(yamlController, 'insertTocEntry');
-		markdownQuickPick();
+		markdownQuickPick(context);
 		await sleep(sleepTime);
 		expect(spy).to.have.been.called();
 		stubShowQuickPick.restore();
@@ -544,7 +596,7 @@ suite('Quick Pick Menu Controller', () => {
 		};
 		stubShowQuickPick.onCall(0).resolves(item);
 		const spy = chai.spy.on(yamlController, 'insertTocEntryWithOptions');
-		markdownQuickPick();
+		markdownQuickPick(context);
 		await sleep(sleepTime);
 		expect(spy).to.have.been.called();
 		stubShowQuickPick.restore();
@@ -572,7 +624,7 @@ suite('Quick Pick Menu Controller', () => {
 		};
 		stubShowQuickPick.onCall(0).resolves(item);
 		const spy = chai.spy.on(yamlController, 'insertExpandableParentNode');
-		markdownQuickPick();
+		markdownQuickPick(context);
 		await sleep(sleepTime);
 		expect(spy).to.have.been.called();
 		stubShowQuickPick.restore();